New scenicplus Docker image
