+++ src/methods/multi_omics/scenicplus/test.sh
@@ -1,7 +1,2 @@
-viash run src/methods/multi_omics/scenicplus/config.vsh.yaml -- --multiomics_atac ./resources_test/grn-benchmark/multiomics_atac.h5ad \
-    --multiomics_rna ./resources_test/grn-benchmark/multiomics_rna.h5ad \
-    --tf_all ./resources/prior/tf_all.csv  \
-    --prediction ./output/scenicplus/prediction.csv \
-    --cell_topic ./output/scenicplus/cell_topic.csv \
-    --scplus_mdata ./output/scenicplus/scplus_mdata.h5mu \
-    --temp_dir ./output/scenicplus/
+viash run src/methods/multi_omics/scenicplus/config.vsh.yaml -- --multiomics_atac ./resources_test/grn-benchmark/multiomics_atac.h5ad --multiomics_rna ./resources_test/grn-benchmark/multiomics_rna.h5ad --tf_all ./resources/prior/tf_all.csv --prediction ./output/scenicplus/prediction.csv --cell_topic ./output/scenicplus/cell_topic.csv --scplus_mdata ./output/scenicplus/scplus_mdata.h5mu --temp_dir ./output/scenicplus/
+#viash run src/methods/multi_omics/scenicplus/config.vsh.yaml -p docker -- ---setup build
